Imported from resin-4.0.65-src.tar.gz.

+++ modules/resin/src/com/caucho/server/http/AbstractResponseStream.java
@@ -234,7 +234,7 @@ public abstract class AbstractResponseStream
   /**
    * Returns the written content length
    */
-  public int getContentLength()
+  public long getContentLength()
   {
     return 0;
   }
